refactor(test): split ResponseEndpointSpec from RequestEndpointSpec
